+++ examples/npt_equilibration_at_multiple_lambdas.py
@@ -9,12 +9,11 @@ from testsystems.relative import hif2a_ligand_pair
 from md.builders import build_water_system
 from md.minimizer import minimize_host_4d
 
-from fe.topology import SingleTopology
 from fe.free_energy import AbsoluteFreeEnergy
 
-from barostat.ensembles import PotentialEnergyModel, NPTEnsemble
-from barostat.moves import MonteCarloBarostat, CoordsAndBox
-from barostat.utils import get_group_indices, compute_box_volume
+from md.ensembles import PotentialEnergyModel, NPTEnsemble
+from md.barostat.moves import MonteCarloBarostat, CoordsAndBox
+from md.barostat.utils import get_group_indices, compute_box_volume
 
 from timemachine.lib import custom_ops, LangevinIntegrator
 from timemachine.constants import BOLTZ
